+++ AddOns/Syndicator/Core/Config.lua
@@ -12,6 +12,11 @@ Syndicator.Config.Options = {
   TOOLTIPS_FACTION_ONLY = "tooltips_faction_only",
   TOOLTIPS_CHARACTER_LIMIT = "tooltips_character_limit",
   TOOLTIPS_SORT_BY_NAME = "tooltips_sort_by_name",
+  SHOW_BLANK_LINE_BEFORE_INVENTORY = "show_blank_line_before_inventory",
+  SHOW_TOTAL_LINE_AFTER_CHARACTERS = "show_total_line_after_characters",
+
+  AUCTION_VALUE_SOURCE = "auction_value_source",
+  NO_AUCTION_VALUE_SOURCE = "no_auction_value_source",
 
   DEBUG = "debug",
   DEBUG_TIMERS = "debug_timers",
@@ -28,6 +33,10 @@ Syndicator.Config.Defaults = {
   [Syndicator.Config.Options.TOOLTIPS_FACTION_ONLY] = false,
   [Syndicator.Config.Options.TOOLTIPS_CHARACTER_LIMIT] = 4,
   [Syndicator.Config.Options.SHOW_CHARACTER_RACE_ICONS] = true,
+  [Syndicator.Config.Options.SHOW_BLANK_LINE_BEFORE_INVENTORY] = false,
+  [Syndicator.Config.Options.SHOW_TOTAL_LINE_AFTER_CHARACTERS] = false,
+  [Syndicator.Config.Options.AUCTION_VALUE_SOURCE] = "", -- "auctionator" ,"oribos-region", "oribos-realm", "tradeskillmaster-dbrecent", etc.
+  [Syndicator.Config.Options.NO_AUCTION_VALUE_SOURCE] = false,
 
   [Syndicator.Config.Options.DEBUG] = false,
   [Syndicator.Config.Options.DEBUG_TIMERS] = false,