+++ AddOns/Syndicator/Core/Utilities.lua
@@ -32,10 +32,14 @@ function Syndicator.Utilities.GetCharacterFullName()
 end
 
 if Syndicator.Constants.IsClassic then
-  local tooltip = CreateFrame("GameTooltip", "BaganatorUtilitiesScanTooltip", nil, "GameTooltipTemplate")
+  local tooltip = CreateFrame("GameTooltip", "SyndicatorUtilitiesScanTooltip", nil, "GameTooltipTemplate")
+  Syndicator.Utilities.ScanningTooltip = tooltip
   tooltip:SetOwner(WorldFrame, "ANCHOR_NONE")
 
   function Syndicator.Utilities.DumpClassicTooltip(tooltipSetter)
+    if Syndicator.Constants.IsBrokenTooltipScanning then
+      return {lines = {}}
+    end
     tooltip:SetOwner(WorldFrame, "ANCHOR_NONE")
     tooltipSetter(tooltip)