+++ AddOns/Syndicator/Search/Utilities.lua
@@ -13,30 +13,23 @@ function Syndicator.Search.GetExpansionInfo(itemID)
   if ItemVersion then
     local itemVersionDetails = ItemVersion.API:getItemVersion(itemID, true)
     if itemVersionDetails then
-      return itemVersionDetails.major, itemVersionDetails.minor, itemVersionDetails.patch
+      return itemVersionDetails.major
     end
   end
   if ATTC and ATTC.SearchForField then
-    local attResults = ATTC.SearchForField("itemID", itemID)
-    local id
-    for _, result in ipairs(attResults) do
-      if result.awp then
-        id = result.awp
-        break
-      end
-    end
-    if not id and #attResults > 0 then
-      local parent = attResults[1]
-      while parent and parent.awp == nil do -- awp: short for added with patch
+    local results = ATTC.SearchForField("itemID", itemID)
+    if #results > 0 then
+      local parent = results[1]
+      local xpac
+      while parent and not xpac do
+        if parent.expansionID then
+          xpac = math.floor(parent.expansionID)
+        end
         parent = parent.parent
       end
-      id = parent and parent.awp
-    end
-    if id then
-      local major = math.floor(id / 10000)
-      local minor = math.floor((id % 10000) / 100)
-      local patch = math.floor(id % 100)
-      return major, minor, patch
+      if xpac then
+        return xpac
+      end
     end
   end
 end
@@ -92,7 +85,6 @@ if Syndicator.Constants.IsRetail then
       return cache[itemLink]
     end
 
-    local start = debugprofilestop()
     local playerActor = modelScene:GetPlayerActor()
     local invType = select(4, C_Item.GetItemInfoInstant(itemLink))
     local slot = SlotMap[invType]
@@ -119,4 +111,101 @@ else
   function Syndicator.Search.RecoverTransmogInfo(itemLink)
     return nil
   end
+end
+
+do
+  local priceSources = {}
+  priceSources["none"] = {
+    func = function(itemLink)
+      return -2 -- special value, says everything matches this price
+    end,
+    priority = 1000,
+    validation = function() return true end
+  }
+  priceSources["auctionator-latest"] = {
+    func = function(itemLink)
+      return Auctionator.API.v1.GetAuctionPriceByItemLink("Syndicator", itemLink)
+    end,
+    priority = 100,
+    validation = function() return Auctionator ~= nil end
+  }
+  priceSources["tradeskillmaster-dbmarket"] = {
+    func = function(itemLink)
+      return TSM_API.GetCustomPriceValue("dbmarket", TSM_API.ToItemString(itemLink))
+    end,
+    priority = 30,
+    validation = function() return TSM_API ~= nil end
+  }
+  priceSources["tradeskillmaster-dbrecent"] = {
+    func = function(itemLink)
+      return TSM_API.GetCustomPriceValue("dbrecent", TSM_API.ToItemString(itemLink))
+    end,
+    validation = function() return TSM_API ~= nil end
+  }
+  priceSources["tradeskillmaster-dbregionmarketavg"] = {
+    func = function(itemLink)
+      return TSM_API.GetCustomPriceValue("dbregionmarketavg", TSM_API.ToItemString(itemLink))
+    end,
+    validation = function() return TSM_API ~= nil end
+  }
+  priceSources["tradeskillmaster-dbregionsaleavg"] = {
+    func = function(itemLink)
+      return TSM_API.GetCustomPriceValue("dbregionmarketavg", TSM_API.ToItemString(itemLink))
+    end,
+    validation = function() return TSM_API ~= nil end
+  }
+  priceSources["undermineexchange-realm"] = {
+    func = function(itemLink)
+      local o = {}
+      OEMarketInfo(itemLink, o)
+      return o["market"] or o["region"]
+    end,
+    priority = 50,
+    validation = function() return OEMarketInfo ~= nil end
+  }
+  priceSources["undermineexchange-region"] = {
+    func = function(itemLink)
+      local o = {}
+      OEMarketInfo(itemLink, o)
+      return o["region"]
+    end,
+    validation = function() return OEMarketInfo ~= nil end
+  }
+  priceSources["recrystallize"] = {
+    func = function(itemLink)
+      return (RECrystallize_PriceCheck(itemLink))
+    end,
+    validation = function() return RECrystallize_PriceCheck ~= nil end,
+    priority = 40,
+  }
+
+  local frame = CreateFrame("Frame")
+  frame:RegisterEvent("PLAYER_LOGIN")
+  frame:SetScript("OnEvent", function()
+    local key = Syndicator.Config.Get(Syndicator.Config.Options.AUCTION_VALUE_SOURCE)
+    local current = priceSources[key]
+    if not current or not current.validation() or (key == "none" and not Syndicator.Config.Get(Syndicator.Config.Options.NO_AUCTION_VALUE_SOURCE)) then
+      local options = {}
+      for key, details in pairs(priceSources) do
+        if details.priority and details.validation() then
+          details.key = key
+          table.insert(options, details)
+        end
+      end
+      table.sort(options, function(a, b) return a.priority < b.priority end)
+      if #options > 0 then
+        current = options[1]
+        Syndicator.Config.Set(Syndicator.Config.Options.AUCTION_VALUE_SOURCE, options[1].key)
+      else
+        error("unexpected missing \"none\" price source")
+      end
+    end
+    Syndicator.Search.GetAuctionValue = current.func
+  end)
+  Syndicator.CallbackRegistry:RegisterCallback("AuctionValueSourceChanged", function()
+    local key = Syndicator.Config.Get(Syndicator.Config.Options.AUCTION_VALUE_SOURCE)
+    Syndicator.Config.Set(Syndicator.Config.Options.NO_AUCTION_VALUE_SOURCE, key == "none")
+    local current = priceSources[key]
+    Syndicator.Search.GetAuctionValue = current.func
+  end)
 end